:loudspeaker: Renaming Components
