+++ src/components/Tags.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+export class Tags extends Component {
+  render() {
+    return (
+      <div>
+
+      </div>
+    )
+  }
+}
+
+export default Tags
